Make the interface nice to use with other OS
We now have a mmm.js class to implement stuff that is either based on privates
on the class itself or publics form the OS specifics.
The plan is to implement the same interface for all OS and then use the most
expensive. Then we add stub functions for linux/osx for things like
openProcess (they dont seem to need that)